Refactor initialization of telescope parameters, move to own module
