+++ ctapipe/core/__init__.py
@@ -3,11 +3,12 @@
 Core functionality of ctapipe
 """
 
-from .component import Component, TelescopeComponent, non_abstract_children
+from .component import Component, non_abstract_children
 from .container import Container, DeprecatedField, Field, FieldValidationError, Map
 from .feature_generator import FeatureGenerator
 from .provenance import Provenance, get_module_version
 from .qualityquery import QualityCriteriaError, QualityQuery
+from .telescope_component import TelescopeComponent
 from .tool import Tool, ToolConfigurationError, run_tool
 
 __all__ = [
